// src/app/features/admin-management/project-management/models/user-alloc-dealloc.model.ts
export interface ProjectUserAllocDelloc {
  userId: number;
  projectId: number;
  userName: string;
  userType: string;
  role: string;
  baseLocation: string
  projectStartDate: string;
  projectEndDate: string;
  state: string,
  organization: string
  type: string
  searchValue: string
  districts: [],
  roleCds: [],

}